feat: add PlayerAuthentication concern and apply to FrontendController
Issue #48 対応

Player 認証ロジックを concern として共通化し、コードの重複を防ぐ。
FrontendController で /frontend/ アクセス時に自動的に Player を作成し Cookie を発行。

変更点:
- app/controllers/concerns/player_authentication.rb を新規作成
  - find_or_create_player_from_cookie メソッドを提供
  - Cookie 設定ロジックを共通化
- GraphqlController を concern を使用するようリファクタリング
  - 重複していた Player 認証ロジックを削除
- FrontendController に Player 認証を追加
  - /frontend/ へのアクセス時に Player 作成 + Cookie 発行
  - /frontend/admin へのアクセス時は認証をスキップ

動作確認:
- /frontend/ → Cookie 発行 & Player 作成 ✅
- /frontend/admin → Cookie 発行なし ✅
- GraphQL API → Cookie 発行 & Player 作成 ✅
- 全テスト通過 (20 runs, 98 assertions, 0 failures)
